requirements.txt
streamlit==1.51.0
google-generativeai==0.8.5
moviepy==2.2.1
proglog==0.1.12
